% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accidents.R
\docType{data}
\name{accidents}
\alias{accidents}
\title{Accidents data set.}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{HOUR}{Hoursin miles from fire station}
  \item{ACCIDENTS}{Number of dollars}
}
}
\source{
\url{https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences-Sixth-Edition/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
accidents
}
\description{
A data set concerning damage from accidents
}
\details{
This is a standard data set from the course MATH 4753 taken from the data sets provided by the text book.
}
\keyword{datasets}
