% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicstan.R
\name{basicstan}
\alias{basicstan}
\title{Run a Bayesian model using STAN}
\usage{
basicstan()
}
\value{
An MCMC object
}
\description{
A basic Bayesian model estimating a population mean
}
\details{
Please note the basic structure of the server input output
}
\section{STAN}{

This function will start a STAN instance creating a STAN object which can be viewed using other packages.
}

\examples{
\dontrun{ basicstan()}
}
