% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonjour.R
\name{bonjour}
\alias{bonjour}
\title{"Bonjour !" : Cette fonction est un exemple de fonction graphique avec deux paramètres simples : la couleur du texte et la couleur du fond.}
\usage{
bonjour(couleur = "pink", fond = "limegreen")
}
\arguments{
\item{couleur, }{couleur du texte ("pink" par défaut)}

\item{fond, }{couleur du fond ("limegreen" par défaut)}
}
\value{
Un graphe simple
}
\description{
"Bonjour !" : Cette fonction est un exemple de fonction graphique avec deux paramètres simples : la couleur du texte et la couleur du fond.
}
\examples{
bonjour()
}
