% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon.R
\docType{data}
\name{carbon}
\alias{carbon}
\title{Iron making process.}
\format{
A data frame with 25 rows and 2 variables:
\describe{
  \item{PilotPlant}{Carbon Content \%, y}
  \item{LabFurnace}{Carbon Content \%, x}
}
}
\source{
\url{https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences-Sixth-Edition/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
carbon
}
\description{
A data set concerning a new innovative iron making process.
}
\details{
This is a standard data set from the course MATH 4753 taken from the data sets provided by the text book.
}
\keyword{datasets}
