% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonjour.R
\docType{data}
\name{fruits}
\alias{fruits}
\title{Composition nutritionnelle de 57 composés alimentaires issus du groupe "fruits" défini dans les données Ciqual de 2020.}
\description{
Composition nutritionnelle de 57 composés alimentaires issus du groupe "fruits" défini dans les données Ciqual de 2020.
}
\references{
\url{https://ciqual.anses.fr/}
}
\keyword{data}
