% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intro2r.R
\docType{package}
\name{intro2r-package}
\alias{intro2r}
\alias{intro2r-package}
\title{Package pour la formation "Introduction au Logiciel R"
Contient les vignettes (les slides de cours), le jeu de données "fruits", le jeu de données "nutrimenu"  et les tutoriels.}
\description{
Ce package contient des tutoriels, les présentations sous forme de vignettes, ainsi qu'un exemple de données.
}
\author{
\strong{Maintainer}: Vincent Guillemot \email{vincent.guillemot@pasteur.fr}

}
\keyword{internal}
