% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lichen.R
\docType{data}
\name{lichen}
\alias{lichen}
\title{Radiation in lichen}
\format{
A data frame with 9 rows and 3 variables:
\describe{
  \item{LOCATION}{Location of lichen}
  \item{LOGCES}{Log of Ces}
  \item{CESIUM}{Ces}
}
}
\source{
\url{https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences-Sixth-Edition/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
lichen
}
\description{
A data set concerning radiation in lichen
}
\details{
This is a standard data set from the course MATH 4753 taken from the data sets provided by the 6th edition of the text book.
}
\keyword{datasets}
