% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybart.R
\name{mybart}
\alias{mybart}
\title{Hypothesis testing Introduction}
\usage{
mybart(n = 10, r = 2)
}
\arguments{
\item{n}{Number of trials}

\item{r}{Number of Blue upper/lwer tail bars}
}
\value{
Barplot
}
\description{
Hypothesis testing Introduction
}
\examples{
mybart()
}
