% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial Simulation}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of iterations}

\item{n}{Number of Bernoulli trials}

\item{p}{Probability of a success}
}
\value{
Barplot and table of relative frequencies
}
\description{
Shows how to simulate a Binomial
}
\details{
More details about the function
}
\examples{
mybin() # Use default values
}
