% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot}
\alias{myboot}
\title{Bootstrap function}
\usage{
myboot(iter = 10000, x, fun = "mean", alpha = 0.05, ...)
}
\arguments{
\item{iter}{Number of iterations}

\item{x}{Sample data}

\item{fun}{function to use}

\item{alpha}{confidence level}

\item{...}{other arguments}
}
\value{
A fancy bootstrap histogram
}
\description{
Bootstrap function
}
\examples{
\dontrun{
sam=rnorm(20,mean=10,sd=2)
myboot(iter=10000, x=sam, fun="mean", alpha=0.05, xlab="mean")
}
}
