% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Central Limit Theorem Uniform}
\usage{
mycltu(n = 3, iter = 10000, a = 0, b = 10)
}
\arguments{
\item{n}{sample size of each iteration}

\item{iter}{iterations}

\item{a}{lower parameter (default = 0)}

\item{b}{upper parameter (default = 10)}
}
\value{
A histogram of a sample mean
}
\description{
Central Limit Theorem Uniform
}
\examples{
\dontrun{
mycltu(n=3, iter=10000, a = 0, b = 10)
}
}
