% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstfun.R
\name{myfirstfun}
\alias{myfirstfun}
\title{Very first function}
\usage{
myfirstfun(x, ...)
}
\arguments{
\item{x}{This is the name of a vector}

\item{...}{extra arguments which will be sent to the corresponding ... in this case the plot function}
}
\value{
A list of x and y data in vector form and a matrix of the same as well as  a plot of the data
}
\description{
This instructs the student on the basics of function making
}
\details{
Welcome to STATS at OU -- we will be learning R from the ground up. RAR man!
}
\section{Making Functions}{


Learn every part of this simple but important function.

\enumerate{
\item Where does the name of the function go?
\item What does the ellipsis do?
\item How long is the object \code{y} in relation to \code{x}?
\item Where do you find more information about \code{plot}?
\item What is a list?
\item what does the function \code{return} do?
}
}

\examples{
####### some examples ####
data <- 1:40
myfirstfun(x = data, pch=21, bg="Blue", cex = 2.4) # x data and extra options to send to plot
}
\seealso{
\code{\link{plot}}
}
