% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhyper.R
\name{myhyper}
\alias{myhyper}
\title{Hypergeometric Simulation}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{Number of iterations}

\item{N}{Number of columns/population size}

\item{r}{Number of successes}

\item{n}{Sample size}
}
\value{
A fancy colored graph with possible successes
}
\description{
Hypergeometric Simulation
}
\examples{
\dontrun{
myhyper(iter=100,n=5, N=20,r=12)
}
}
