% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymode.R
\name{mymode}
\alias{mymode}
\title{The mode}
\usage{
mymode(x)
}
\arguments{
\item{x}{vector of data}
}
\value{
the mode of the data (most frequent)
}
\description{
The mode
}
\section{Origin}{
 Taken from a web page originally and then adapted \url{https://www.tutorialspoint.com/r/r_mean_median_mode.htm}
}

\examples{
\dontrun{
mymode(ddt$LENGTH)}
}
