% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myttest.R
\name{mynewf}
\alias{mynewf}
\title{Another function}
\usage{
mynewf(x)
}
\arguments{
\item{x}{vector}
}
\value{
plot
}
\description{
Another function
}
\details{
You may add as many functions as you wish in this file and document them in Roxygen
}
\examples{
x<-1:10;mynewf(x)
}
