% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfun2.R, R/myplot.R
\name{myplot}
\alias{myplot}
\title{My special function}
\usage{
myplot(x)

myplot(x)
}
\arguments{
\item{x}{The diameter}
}
\value{
y hat value

Response ie estimated Height
}
\description{
My special function

A function for plotting a quadratic
}
\examples{
\dontrun{myplot(x=13)}
\dontrun{quad.lm<-lm(y~x,data=df); myplot(x=15)}
}
