% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myreadxl.R
\name{myreadxl}
\alias{myreadxl}
\title{Read all xl files}
\usage{
myreadxl(dird = "D:/MATH4773-5773/DATA/Excel/")
}
\arguments{
\item{dird}{this is the directory structure as a string}
}
\value{
a named list of the files read
}
\description{
This function will read a collection of xls files into R's workplace
}
\details{
This will greatly reduce the amount of work needed to perform analyses since the reading of files can be time consuming.
}
\examples{
\dontrun{
v<-myreadxl()
}
}
