% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myttest.R
\name{myttest}
\alias{myttest}
\title{A constructor function for t tests}
\usage{
myttest(x, y, paired = FALSE, alpha = 0.05)
}
\arguments{
\item{x}{vector of data}

\item{y}{vector of data}

\item{paired}{TRUE or FALSE, FALSE by default}

\item{alpha}{alpha level between 0,1}
}
\value{
list containing the data, and t test object
}
\description{
This is an example of what can be done with S3 OOP
}
\details{
The constructor function makes a list that can be worked and further processed with an appropriate method(s)
}
\examples{
myttest(x=rnorm(30,10,12), y=rnorm(40, 7, 10))
}
\seealso{
\code{\link{t.test}} for more information about t tests
}
\keyword{NHST}
\keyword{t.test}
\keyword{testing}
\keyword{ttest}
