% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojuice.R
\docType{data}
\name{ojuice}
\alias{ojuice}
\title{Orange Juice.}
\format{
A data frame with 24 rows and 2 variables:
\describe{
  \item{SweetIndex}{Higher the index the sweeter the juice}
  \item{Pectin}{Water soluble Pectin in ppm}
  \item{Run}{Integer for run number}
}
}
\source{
\url{https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences-Sixth-Edition/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
ojuice
}
\description{
A data set concerning Orange Juice
}
\details{
This is a standard data set from the course MATH 4753 taken from the data sets provided by the text book.
}
\keyword{datasets}
