% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmytt.R
\name{plot.mytt}
\alias{plot.mytt}
\title{Method for plotting class mytt}
\usage{
\method{plot}{mytt}(x, ...)
}
\arguments{
\item{x}{object class mytt}

\item{...}{extra options to be sent to method}
}
\value{
A boxplot of the two samples using \code{ggplot()}
}
\description{
Makes a boxplot using an object of class mytt
}
\details{
The function \code{myttest()} produces a list of class mytt, which has three components. These are sourced from the object and used in \code{plot.mytt()}
}
\examples{
l<-myttest(x=rnorm(30), y=rnorm(40,0.5));plot(l)
}
