% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Rttest}
\alias{print.Rttest}
\title{Print method for class Rttest}
\usage{
\method{print}{Rttest}(x, ...)
}
\arguments{
\item{x}{data from constructor function}

\item{...}{table variables}
}
\value{
a table with the data and the alpha value, confidence intervals and p-value.
}
\description{
Print method for class Rttest
}
\examples{
\dontrun{print(x)}
}
\author{
Jordan Daugherty (2022)
}
