% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmv.R
\name{readMV}
\alias{readMV}
\title{Read all DAT data from JW}
\usage{
readMV(dird = "D:\\\\MULTIVARIATE-OU - 2021\\\\DATA\\\\")
}
\arguments{
\item{dird}{Directory to the data folder with separator on the end}
}
\value{
A named list of all the data
}
\description{
Read all DAT data from JW
}
\examples{
\dontrun(v <- readMV())
}
