% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdfiles.R
\name{rmdfiles}
\alias{rmdfiles}
\title{Links to files in directory}
\usage{
rmdfiles(rootf, pack = "Intro2R")
}
\arguments{
\item{rootf}{this is the folder in the root of the package where files are located}

\item{pack}{this is the package name, Intro2R is default}
}
\value{
rmd code for setting up link
}
\description{
This will place a link in the RMD document to files in the Intro2R package
}
\details{
This makes the construction of links easier
}
\examples{
\dontrun{
`r rmdfiles(rootf="Lab2")`
}
}
