% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterhist.R
\name{scatterhist}
\alias{scatterhist}
\title{Scatter histogram From Lab1}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{X variable}

\item{y}{Y variable}

\item{xlab}{x label}

\item{ylab}{y label}
}
\value{
A fancy scattered histogram plot
}
\description{
Scatter histogram From Lab1
}
\examples{
\dontrun{
with(ddt, scatterhist(LENGTH,WEIGHT, xlab="LENGTH"))
}
}
