% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\docType{data}
\name{seed}
\alias{seed}
\title{Germination of seeds}
\format{
A data frame with 7 rows and 2 variables:
\describe{
  \item{CHANGE}{Change in free energy, \code{y}}
  \item{TEMP}{Temperature}
}
}
\source{
\url{https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences-Sixth-Edition/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
seed
}
\description{
A data set concerning seed germination
}
\details{
This is a standard data set from the course MATH 4753 taken from the data sets provided by the 5th edition of the text book.
}
\keyword{datasets}
