% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spruce.R
\docType{data}
\name{spruce}
\alias{spruce}
\title{Spruce data set.}
\format{
A data frame with 36 rows and 2 variables:
\describe{
  \item{BHDiameter}{Breast height diameter of the trees}
  \item{Height}{Height of the trees}
}
}
\source{
\url{https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences-Sixth-Edition/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
spruce
}
\description{
A data set concerning Spruce trees
}
\details{
This is a standard data set from the course MATH 4753 taken from the data sets provided by the text book.
}
\keyword{datasets}
