% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicstan.R
\name{basicstan}
\alias{basicstan}
\title{Basic Stan}
\usage{
basicstan(chains = 3, warmup = 1000, iter = 5000, html = FALSE)
}
\arguments{
\item{chains}{The number of MCMC chains of simulation}

\item{warmup}{How many iterations to throw away}

\item{iter}{The total number of iterations}

\item{html}{TRUE if you want to see a browser interface}
}
\value{
An MCMC object
}
\description{
A basic Bayesian model estimating a population mean
}
\details{
Please check the STAN site for documentation
}
\section{STAN}{

This function will start a STAN instance creating a STAN object which can be viewed using other packages.
}

\examples{
\dontrun{ basicstan()}
}
