% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT data set}
\format{
A data frame with 144 rows and 5 variables:
\describe{
  \item{LENGTH}{length in cm}
  \item{WEIGHT}{weight of the fish, in gms}
  \item{MILE}{mile where collected}
  \item{RIVER}{river name}
  \item{SPECIES}{fish species}
  \item{DDT}{DDT concentration}
}
}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
ddt
}
\description{
A data set of fish collected in the tennessee river
}
\keyword{datasets}
