% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drill.R
\docType{data}
\name{drill}
\alias{drill}
\title{Hydraulic drilling experiment.}
\format{
A data frame with 17 rows and 2 variables:
\describe{
  \item{DEPTH}{Depth at which drilling occurs, feet}
  \item{TIME}{Time to drill 5 feet, minutes}
}
}
\source{
\url{https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences-Sixth-Edition/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
drill
}
\description{
A data set concerning hydraulic drilling in rock
}
\details{
The time  \code{TIME} to drill five feet in rock is measured at different depths \code{DEPTH}
}
\keyword{datasets}
