% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymode.R
\name{mymode}
\alias{mymode}
\title{Function to find the mode}
\usage{
mymode(v)
}
\arguments{
\item{v}{vector of data}
}
\value{
the mode of the data (most frequent)
}
\description{
Function to find the mode
}
\examples{
\dontrun{
mymode(ddt$LENGTH)}
}
