% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot.R
\name{myplot}
\alias{myplot}
\title{A function for plotting a quadratic}
\usage{
myplot(x, q)
}
\arguments{
\item{x}{The diameter}

\item{q}{a quadratic lm object}
}
\value{
Response ie estimated Height
}
\description{
A basic function
}
\examples{
\dontrun{quad.lm<-lm(y~x,data=df); myplot(x=15, q = quad.lm)}
}
