% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot.R
\name{myplot}
\alias{myplot}
\title{A function for plotting a quadratic}
\usage{
myplot(x)
}
\arguments{
\item{x}{The diameter}
}
\value{
Response ie estimated Height
}
\description{
A function for plotting a quadratic
}
\examples{
\dontrun{quad.lm<-lm(y~x,data=df); myplot(x=15)}
}
